+++ src/build-tasks/tasks/copy-to-s3-task.ts
@@ -1,19 +1,24 @@
 import path from 'path';
+import { existsSync } from 'fs';
 import { ConsoleUtil } from '../../../src/console-util';
+import { OrgFormationError } from '../../../src/org-formation-error';
 import { IBuildTaskConfiguration, IBuildTask } from '~build-tasks/build-configuration';
 import { IOrganizationBinding } from '~parser/parser';
 import { IBuildTaskProvider, BuildTaskProvider } from '~build-tasks/build-task-provider';
 import { IPerformTasksCommandArgs } from '~commands/index';
 import { IS3CopyCommandArgs, S3CopyCommand } from '~commands/s3copy/s3copy';
+import { Validator } from '~parser/validator';
 
 export class CopyToS3TaskProvider implements IBuildTaskProvider<ICopyToS3TaskConfiguration> {
     public type = 'copy-to-s3';
 
     createTask(config: ICopyToS3TaskConfiguration, command: IPerformTasksCommandArgs): IBuildTask {
+        CopyToS3TaskProvider.validateConfig(config);
+
         return {
             type: config.Type,
             name: config.LogicalName,
-            physicalIdForCleanup: config.LogicalName,
+            // physicalIdForCleanup: config.LogicalName,
             childTasks: [],
             isDependency: BuildTaskProvider.createIsDependency(config),
             perform: async (): Promise<void> => {
@@ -31,6 +36,7 @@ export class CopyToS3TaskProvider implements IBuildTaskProvider<ICopyToS3TaskCon
                     failedTolerance:config.FailedTaskTolerance,
                     maxConcurrent: config.MaxConcurrentTasks,
                     organizationBinding: config.OrganizationBinding,
+                    taskRoleName: config.TaskRoleName,
                 };
 
                 await S3CopyCommand.Perform(updateSlsCommand);
@@ -38,17 +44,16 @@ export class CopyToS3TaskProvider implements IBuildTaskProvider<ICopyToS3TaskCon
         };
     }
 
-    createTaskForValidation(/* config: ICopyToS3TaskConfiguration, command: IPerformTasksCommandArgs*/): IBuildTask | undefined {
-        return undefined;
-        // return {
-        //     type: config.Type,
-        //     name: config.LogicalName,
-        //     childTasks: [],
-        //     isDependency: (): boolean => false,
-        //     perform: async (): Promise<void> => {
-
-        //     },
-        // };
+    createTaskForValidation(config: ICopyToS3TaskConfiguration): IBuildTask | undefined {
+        return {
+            type: config.Type,
+            name: config.LogicalName,
+            childTasks: [],
+            isDependency: (): boolean => false,
+            perform: async (): Promise<void> => {
+                CopyToS3TaskProvider.validateConfig(config);
+            },
+        };
     }
 
     createTaskForCleanup(/* logicalId: string, physicalId: string, command: IPerformTasksCommandArgs*/): IBuildTask | undefined {
@@ -64,6 +69,25 @@ export class CopyToS3TaskProvider implements IBuildTaskProvider<ICopyToS3TaskCon
         //     },
         // };
     }
+
+
+    static validateConfig(config: ICopyToS3TaskConfiguration): void {
+        if (!config.LocalPath) {
+            throw new OrgFormationError(`task ${config.LogicalName} does not have required attribute LocalPath`);
+        }
+        if (!config.RemotePath) {
+            throw new OrgFormationError(`task ${config.LogicalName} does not have required attribute RemotePath`);
+        }
+
+        const dir = path.dirname(config.FilePath);
+        const localPath = path.join(dir, config.LocalPath);
+
+        if (!existsSync(localPath)) {
+            throw new OrgFormationError(`task ${config.LogicalName} cannot find path ${config.FilePath}`);
+        }
+
+        Validator.ValidateOrganizationBinding(config.OrganizationBinding, config.LogicalName);
+    }
 }
 
 
@@ -74,4 +98,5 @@ export interface ICopyToS3TaskConfiguration extends IBuildTaskConfiguration {
     OrganizationBinding: IOrganizationBinding;
     MaxConcurrentTasks?: number;
     FailedTaskTolerance?: number;
+    TaskRoleName?: string;
 }