+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 All notable changes to aws organization formation will be documented in this file.
 
 **not released**
-- added --organization-file to perform-tasks and validate-tasks command, can be used to specify the organization to be used when evaluating bindings (without having an update-organization task)
+- Added update-serverless.com task to perform-tasks
+- Added copy-to-s3 task to perform-tasks
+- Added --organization-file to perform-tasks and validate-tasks command, can be used to specify the organization to be used when evaluating bindings (without having an update-organization task)
+- Added TaskRoleName to update-stacks, update-serverless.com & copy-to-s3 tasks to support least custom role for cross account access
+- Added CloudFormationRoleName to update-stacks to pass specific role to CloudFormation
 
 **version 0.9.3**
 - Added flag --perform-cleanup to perform-tasks to automatically delete stacks removed from tasks file